+++ src/Core/Statistic/Statistic.Domain/Statistics/CreateExpired/ICreateExpiredStatisticIn.cs
@@ -4,6 +4,6 @@ public interface ICreateExpiredStatisticIn
 {
     public Guid ItemId { get; set; }    
     public Guid UserId { get; set; }
-    public float ItemWeight { get; set; }
+    public double ItemWeight { get; set; }
     
 }
